requirements.txt
img2pdf==0.4.0
selenium==3.141.0
Pillow==7.2.0
